agent/requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0
pillow>=10.0.0

# Deep Learning
torch>=2.0.0
transformers>=4.35.0

# LLM & Agent
openai>=1.0.0
langchain>=0.1.0
requests>=2.31.0

# Document Processing
python-docx>=0.8.11
pyyaml>=6.0

# Progress Bar
tqdm>=4.65.0

# Environment
python-dotenv>=1.0.0
